requirements.txt
alembic==0.9.7
amqp==1.4.9
aniso8601==2.0.0
anyjson==0.3.3
backports.shutil-get-terminal-size==1.0.0
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.25
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
decorator==4.2.1
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.21.0
Faker==0.8.10
Flask==0.12.2
Flask-Caching==1.3.3
Flask-Flash==1.6.7
Flask-HTTPAuth==3.2.3
Flask-JSGlue==0.3.1
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SocketIO==2.9.3
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
greenlet==0.4.12
idna==2.6
inflect==0.2.5
ipaddress==1.0.19
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
kombu==3.0.37
lxml==4.1.1
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
meld3==1.0.2
numpy==1.14.0
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
python-engineio==2.0.2
python-socketio==1.8.4
-e git+https://github.com/ocervell/pytube@b79790ab40bd356f3efa495b2c64782ff6234ffb#egg=pytube
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
scandir==1.6
-e git+https://github.com/ocervell/scrapez.git@bcb780dfb5129c2fbdf445a05768f063d19b64d5#egg=scrapez
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.1
supervisor==3.3.3
text-unidecode==1.1
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.14.1
win-unicode-console==0.5
